Fix the linux build.

+++ libcrocofixutility_tests/number_of_digits_tests.cpp
@@ -25,27 +25,27 @@ TEST_CASE("number_of_digits", "[number_of_digits]") {
         REQUIRE(number_of_digits(99'999'999) == 8);
         REQUIRE(number_of_digits(100'000'000) == 9);
         REQUIRE(number_of_digits(999'999'999) == 9);
-        REQUIRE(number_of_digits(1'000'000'000) == 10);
-        REQUIRE(number_of_digits(9'999'999'999ull) == 10);
-        REQUIRE(number_of_digits(10'000'000'000ull) == 11);
-        REQUIRE(number_of_digits(99'999'999'999ull) == 11);
-        REQUIRE(number_of_digits(100'000'000'000ull) == 12);
-        REQUIRE(number_of_digits(999'999'999'999ull) == 12);
-        REQUIRE(number_of_digits(1'000'000'000'000ull) == 13);
-        REQUIRE(number_of_digits(9'999'999'999'999ull) == 13);
-        REQUIRE(number_of_digits(10'000'000'000'000ull) == 14);
-        REQUIRE(number_of_digits(99'999'999'999'999ull) == 14);
-        REQUIRE(number_of_digits(100'000'000'000'000ull) == 15);
-        REQUIRE(number_of_digits(999'999'999'999'999ull) == 15);
-        REQUIRE(number_of_digits(1'000'000'000'000'000ull) == 16);
-        REQUIRE(number_of_digits(9'999'999'999'999'999ull) == 16);
-        REQUIRE(number_of_digits(10'000'000'000'000'000ull) == 17);
-        REQUIRE(number_of_digits(99'999'999'999'999'999ull) == 17);
-        REQUIRE(number_of_digits(100'000'000'000'000'000ull) == 18);
-        REQUIRE(number_of_digits(999'999'999'999'999'999ull) == 18);
-        REQUIRE(number_of_digits(1'000'000'000'000'000'000ull) == 19);
-        REQUIRE(number_of_digits(9'999'999'999'999'999'999ull) == 19);
-        REQUIRE(number_of_digits(10'000'000'000'000'000'000ull) == 20);
+        REQUIRE(number_of_digits(uint64_t(1'000'000'000)) == 10);
+        REQUIRE(number_of_digits(uint64_t(9'999'999'999)) == 10);
+        REQUIRE(number_of_digits(uint64_t(10'000'000'000)) == 11);
+        REQUIRE(number_of_digits(uint64_t(99'999'999'999)) == 11);
+        REQUIRE(number_of_digits(uint64_t(100'000'000'000)) == 12);
+        REQUIRE(number_of_digits(uint64_t(999'999'999'999)) == 12);
+        REQUIRE(number_of_digits(uint64_t(1'000'000'000'000)) == 13);
+        REQUIRE(number_of_digits(uint64_t(9'999'999'999'999)) == 13);
+        REQUIRE(number_of_digits(uint64_t(10'000'000'000'000)) == 14);
+        REQUIRE(number_of_digits(uint64_t(99'999'999'999'999)) == 14);
+        REQUIRE(number_of_digits(uint64_t(100'000'000'000'000)) == 15);
+        REQUIRE(number_of_digits(uint64_t(999'999'999'999'999)) == 15);
+        REQUIRE(number_of_digits(uint64_t(1'000'000'000'000'000)) == 16);
+        REQUIRE(number_of_digits(uint64_t(9'999'999'999'999'999)) == 16);
+        REQUIRE(number_of_digits(uint64_t(10'000'000'000'000'000)) == 17);
+        REQUIRE(number_of_digits(uint64_t(99'999'999'999'999'999)) == 17);
+        REQUIRE(number_of_digits(uint64_t(100'000'000'000'000'000)) == 18);
+        REQUIRE(number_of_digits(uint64_t(999'999'999'999'999'999)) == 18);
+        REQUIRE(number_of_digits(uint64_t(1'000'000'000'000'000'000)) == 19);
+        REQUIRE(number_of_digits(uint64_t(9'999'999'999'999'999'999ul)) == 19);
+        REQUIRE(number_of_digits(uint64_t(10'000'000'000'000'000'000ul)) == 20);
     }
 
     SECTION("number of digits in negative integers") {
